+++ apps/api/src/modules/auth/dto/request/signin-user.dto.ts
@@ -1,7 +1,6 @@
 import { IsNotEmpty, IsString } from 'class-validator';
 
 export class SignInUserDto {
-
   @IsString()
   @IsNotEmpty()
   username: string;
