+++ apps/api/src/modules/auth/guards/cookie.guard.ts
@@ -2,6 +2,4 @@ import { AuthGuard } from '@nestjs/passport';
 import { Injectable } from '@nestjs/common';
 
 @Injectable()
-export class CookieGuard extends AuthGuard('cookie') {
-
-}
+export class CookieGuard extends AuthGuard('cookie') {}
