+++ apps/api/src/modules/auth/user-session.entity.ts
@@ -1,9 +1,14 @@
-import { Column, Entity, JoinColumn, OneToOne, PrimaryGeneratedColumn } from 'typeorm';
+import {
+  Column,
+  Entity,
+  JoinColumn,
+  OneToOne,
+  PrimaryGeneratedColumn,
+} from 'typeorm';
 import { User } from '../user/user.entity';
 
 @Entity()
 export class UserSession {
-
   @PrimaryGeneratedColumn('uuid')
   id: string;
